Add paunch container shutdown service
Since the containers launched by paunch do not have any integration
with systemd, we need to have a service that will gracefully shutdown
the services prior to system reboots/poweroffs. This change adds a
single service that runs only when the system is shutdown to loop
through all the containers launched by pauch and issue a 'docker stop'.

This change should only shutdown containers with the label
managed_by=paunch. Currently we are hard coding the stop timeout in this
service but with recent additions to paunch, we should be able to drop
the stop time and honor the configured stop_grace_period for the
services.

Change-Id: If42a4f43c68c14c5fffdfd9253b68e1eb358e311
Related-Bug: LP#1778913
(cherry picked from commit 73160694b7b21aba641867eafee958351c2c2b61)