Make sure paunch-container-shutdown stops before pacemaker
paunch-container-shutdown has no 'After=pacemaker.service' in the service definition
We want this for two reasons:

A.1)
We want the normal openstack services to be able to stop before the
DB/rabbit so they can run any actions that need to talk to DB/messaging on
shutdown correctly.

A.2)
paunch-container-shutdown stops 10 containers
concurrently at the same time which on some environments creates quite
the load. Doing that and shutting down the HA containers at the same
time is less than desirable.

Change-Id: I669e4ea319268fb478dd523b676b3581a2154cb8
Related-Bug: #1792701