Ensure we actually have /run/netns before starting containers
As /run is cleaned up upon reboot, we lose the /run/netns with each
reboot.

With podman support being added, we must first ensure host directories
exist before starting a container with bind-mounts, as podman will not
create missing trees.

This simple, dummy unit ensures a "placeholder" network namespace is create
before any podman container is started with systemd, ensuring container
bind-mounting the /run/netns location will actually start.

The "placeholder" namespace is then removed, in order to avoid confusion.

It also adds a direct dependency on iproute, in order to ensure we actually
have the command at the right time for the unit to be launched.

Closes-Bug: #1799638
Change-Id: Iec5179470820524926b3f93fb6c5a094bcb4fe0c