Remove Docker support
We won't support Docker in Stein.
If docker is needed on the host, we have an ansible role that will
deploy it but let's not install the rpm anymore by default.

Note: we keep the logic in the systemd service & shutdown script because
we still deploy Docker when Pacemaker is enabled on CentOS7.

Change-Id: Ia3b7e8aab7cc4d211f6f80f25897251a842aca64