paunch-container-shutdown: do not stop containers if podman is enabled
Containers that are managed by Paunch and not Pacemaker, use SystemD for
their lifecycle. So stopping the container like we did, is wrong and we
should rather let SystemD stopping them when the system is asked to
reboot or shutdown.

To do that, paunch-container-shutdown is modified to only stop
containers when running on CentOS7 & Docker. It's the only case we need
to stop the containers (a comment was added to the
paunch-container-shutdown.service).

The expected workflow is the following:

During a shutdown or reboot:

1. All containers managed by Paunch and not by Pacemaker will shutdown.
   depends_on can be used in THT to setup some specific order between
   these containers if needed (already the case for Neutron/OVS).

At this stage, nothing is using RabbitMQ or Galera.

2. paunch-container-shutdown.service shuts down which runs
   paunch-container-shutdown which will stop containers with docker CLI
   if on CentOS7 & Docker enabled otherwise do nothing.

If Podman is used, and some containers managed by Paunch are still
alive, we notify the operator via logger.

3. Pacemaker shuts down, bringing down all containers managed by
   Pacemaker.
4. All containers are supposed to be down and the system can stop or
   restart.

During the start:

1. Pacemaker starts first; bringing up its containers.

2. paunch-container-shutdown.service is started and waits for a
   shutdown.

3. All paunch-managed containers start in the order defined by
   depends_on or at the same time if no order.

Change-Id: Ic9c68ea9d66f895cbb2fa59c3c7682b83176d892
(cherry picked from commit d2ee4a8a24084b54abd95dabd58d554942e70efe)