New script to inject drop-in dependencies into systemd scopes
This patch installs a new script on the host:
paunch-start-podman-container

This patch will be the entrypoint used by Systemd services in
ExecStart parameter.

The script injects a dependency in libpod-conmon*.scopes libpod-*.scopes
so that systemd is forbidden to stop those scopes
automatically until paunch-container-shutdown.service is stopped.
That way, when systemd stops tripleo_*.service, the two scopes will
finish in sequence and paunch dependencies will be respected.

rhbz#1726442
Change-Id: I286479b76e2840ff048f2deb08d475a41e04cbba