[Train] Ensure we don't get a wrong, untested podman version
The bug fixes in podman are managed in a minor number, for instance
podman-1.6.4-15.module+el8.2.0+7290+954fb593.x86_64

The version lock therefore allows to NOT get >1.6.4, but should allow to
get any bugfix versions.

This is linked to https://bugzilla.redhat.com/show_bug.cgi?id=1861777
and should be backported down to train.

Note: this patch differs from the ussuri[1] one since we have to support
rhel-7 based env for train, and it provides an older version of podman.

[1] https://review.rdoproject.org/r/#/q/I518a33c3a2db0565af9205e3dad7e048fc67e9f7

Change-Id: Ibc42c878113896c50f5194d140dbce135f36b7a6